requirement.txt
altgraph==0.17.4
colorama==0.4.6
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
pillow_heif==0.20.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pywin32-ctypes==0.2.3
setuptools==75.6.0
tqdm==4.67.0
